+++ CHANGES.md
@@ -8,6 +8,7 @@ v1.7.0 - YYYY-MM-DD
 - Now use TTF 1.1 or later for font support.
 - Added support for basic compound stream filters for ASCII85Decode support
   (Issue #11)
+- Added support for LZWDecode filters (Issue #11)
 - Fixed a buffer overflow in the (still not enabled) AES-256 code.